Make "control.family" argument for passing to the "inla" function

Make "control.family" argument for passing to the "inla" function

Usage

make_inlami_control.family(
  family_moi,
  error_type = "classical",
  prior.prec.moi = NULL,
  prior.prec.berkson = NULL,
  prior.prec.classical = NULL,
  prior.prec.imp = NULL,
  prior.beta.error = NULL,
  initial.prec.moi = NULL,
  initial.prec.berkson = NULL,
  initial.prec.classical = NULL,
  initial.prec.imp = NULL,
  control.family.moi = NULL,
  control.family.berkson = NULL,
  control.family.classical = NULL,
  control.family.imp = NULL,
  control.family = NULL
)

Arguments

family_moi

a string indicating the likelihood family for the model of interest (the main model).

error_type

type of error (one or more of "classical", "berkson", "missing")

prior.prec.moi

a string containing the parameters for the prior for the precision of the residual term for the model of interest. If this precision should be fixed (to the value given as the initial value), this can be done by specifying "prior.prec.moi = 'fixed'".

prior.prec.berkson

a string containing the parameters for the prior for the precision of the error term for the Berkson error model. If this precision should be fixed (to the value given as the initial value), this can be done by specifying "prior.prec.berkson = 'fixed'".

prior.prec.classical

a string containing the parameters for the prior for the precision of the error term for the classical error model. If this precision should be fixed (to the value given as the initial value), this can be done by specifying "prior.prec.classical = 'fixed'".

prior.prec.imp

a string containing the parameters for the precision of the latent variable x, which is the variable being described in the imputation model. If this precision should be fixed (to the value given as the initial value), this can be done by specifying "prior.prec.imputation = 'fixed'".

prior.beta.error

parameters for the prior for the coefficient of the error prone variable. TODO: Which distribution? Gamma?

initial.prec.moi

the initial value for the precision of the residual term for the model of interest.

initial.prec.berkson

the initial value for the precision of the residual term for the Berkson error term.

initial.prec.classical

the initial value for the precision of the residual term for the classical error term.

initial.prec.imp

the initial value for the precision of the residual term for the latent variable r.

control.family.moi

control.family component for model of interest. Can be specified here using the inla syntax instead of passing the "prior.prec..." and "initial.prec..." arguments, or in the cases when other hyperparameters are needed for the model of interest, see for instance survival models.

control.family.berkson

control.family component Berkson model. Can be specified here using the inla syntax instead of passing the "prior.prec..." and "initial.prec..." arguments. Useful in the cases when more flexibility is needed, for instance if one wants to specify a different prior distribution than Gamma.

control.family.classical

control.family component for classical model. Can be specified here using the inla syntax instead of passing the "prior.prec..." and "initial.prec..." arguments. Useful in the cases when more flexibility is needed, for instance if one wants to specify a different prior distribution than Gamma.

control.family.imp

control.family component for imputation model. Can be specified here using the inla syntax instead of passing the "prior.prec..." and "initial.prec..." arguments. Useful in the cases when more flexibility is needed, for instance if one wants to specify a different prior distribution than Gamma.

control.family

control.family for use in inla (can be provided directly instead of passing the "prior.prec...." and "initial.prec..." arguments. If this is specified, any other "control.family..." or "prior.prec..." arguments provided will be ignored.

Value

the "control.family" argument to be passed to inla, a list of "control.family" arguments for each model in the hierarchical measurement error model.

Examples

make_inlami_control.family(
  formula_moi = simple_moi,
  formula_imp = simple_imp,
  family_moi = "gaussian",
  error_type = c("berkson", "classical"),
  prior.prec.moi = c(10, 9),
  prior.prec.berkson = c(10, 9),
  prior.prec.classical = c(10, 9),
  prior.prec.imp = c(10, 9),
  prior.beta.error = c(0, 1/1000),
  initial.prec.moi = 1,
  initial.prec.berkson = 1,
  initial.prec.classical = 1,
  initial.prec.imp = 1)
#> Error in make_inlami_control.family(formula_moi = simple_moi, formula_imp = simple_imp,     family_moi = "gaussian", error_type = c("berkson", "classical"),     prior.prec.moi = c(10, 9), prior.prec.berkson = c(10, 9),     prior.prec.classical = c(10, 9), prior.prec.imp = c(10, 9),     prior.beta.error = c(0, 1/1000), initial.prec.moi = 1, initial.prec.berkson = 1,     initial.prec.classical = 1, initial.prec.imp = 1): unused arguments (formula_moi = simple_moi, formula_imp = simple_imp)

make_inlami_control.family(
  family_moi = "weibull.surv",
  error_type = c("classical", "missing"),
  control.family.moi =
    list(hyper = list(alpha = list(param = 0.01,
                                   initial = log(1.4),
                                   fixed = FALSE))),
  prior.prec.classical = c(0.5, 0.5),
  prior.prec.imp = c(0.5, 0.5),
  initial.prec.classical = 2.8,
  initial.prec.imp = 1)
#> [[1]]
#> [[1]]$hyper
#> [[1]]$hyper$alpha
#> [[1]]$hyper$alpha$param
#> [1] 0.01
#> 
#> [[1]]$hyper$alpha$initial
#> [1] 0.3364722
#> 
#> [[1]]$hyper$alpha$fixed
#> [1] FALSE
#> 
#> 
#> 
#> 
#> [[2]]
#> [[2]]$hyper
#> [[2]]$hyper$prec
#> [[2]]$hyper$prec$initial
#> [1] 1.029619
#> 
#> [[2]]$hyper$prec$param
#> [1] 0.5 0.5
#> 
#> [[2]]$hyper$prec$fixed
#> [1] FALSE
#> 
#> 
#> 
#> 
#> [[3]]
#> [[3]]$hyper
#> [[3]]$hyper$prec
#> [[3]]$hyper$prec$initial
#> [1] 0
#> 
#> [[3]]$hyper$prec$param
#> [1] 0.5 0.5
#> 
#> [[3]]$hyper$prec$fixed
#> [1] FALSE
#> 
#> 
#> 
#>